+++ freckle-env/doctest/Main.hs
@@ -1,0 +1,11 @@
+module Main
+  ( main
+  )
+where
+
+import Prelude
+
+import Freckle.App.Test.DocTest
+
+main :: IO ()
+main = doctest "library/"